[Refactor] Extract nodeDefSchema from apiSchema
