+++ src/extensions/core/groupNode.ts
@@ -3,12 +3,12 @@ import { type LGraph, LGraphCanvas, LiteGraph } from '@comfyorg/litegraph'
 import { LGraphNode, type NodeId } from '@comfyorg/litegraph/dist/LGraphNode'
 
 import { t } from '@/i18n'
-import type { ComfyNodeDef } from '@/schemas/apiSchema'
 import {
   ComfyLink,
   ComfyNode,
   ComfyWorkflowJSON
 } from '@/schemas/comfyWorkflowSchema'
+import type { ComfyNodeDef } from '@/schemas/nodeDefSchema'
 import { useDialogService } from '@/services/dialogService'
 import { useNodeDefStore } from '@/stores/nodeDefStore'
 import { useToastStore } from '@/stores/toastStore'
