+++ tests-ui/tests/apiTypes.test.ts
@@ -2,7 +2,10 @@
 import fs from 'fs'
 import path from 'path'
 
-import { ComfyNodeDef, validateComfyNodeDef } from '@/schemas/apiSchema'
+import {
+  type ComfyNodeDef,
+  validateComfyNodeDef
+} from '@/schemas/nodeDefSchema'
 
 const EXAMPLE_NODE_DEF: ComfyNodeDef = {
   input: {
